+++ src/main/php/text/ical/Alarm.class.php
@@ -7,7 +7,7 @@ class Alarm implements Object {
   use Alarm\is\Value;
   use Alarm\with\Builder;
 
-  private $description, $trigger, $action;
+  private $description, $trigger, $action, $properties;
 
   /**
    * Write this object
@@ -17,10 +17,13 @@ class Alarm implements Object {
    * @return void
    */
   public function write($out, $name) {
-    $out->object('valarm', [
-      'description' => $this->description,
-      'trigger'    => $this->trigger,
-      'action'     => $this->action
-    ]);
+    $out->object('valarm', array_merge(
+      [
+        'description' => $this->description,
+        'trigger'     => $this->trigger,
+        'action'      => $this->action
+      ],
+      (array)$this->properties
+    ));
   }
 }