+++ ChangeLog.md
@@ -3,6 +3,10 @@ ICal ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Added support for named properties (e.g. `X-MICROSOFT-CDO-*`) to
+  calendars, events and alarms.
+  (@thekid)
+
 ## 0.7.0 / 2016-06-26
 
 * Added support for multiple events inside a calendar